Refactored caching folder
