+++ server/src/cron/caching/Menus/populateMenus.js
@@ -1,10 +1,10 @@
 import * as firebase from "firebase-admin";
-import firestore from "../../config/firebase/firebaseConfig";
+import firestore from "../../../config/firebase/firebaseConfig";
 
 import getOneMenu from "./getOneMenu";
 
-import locations from "../../config/locations";
-import { E_DB_WRITE } from "../../config/constants";
+import locations from "../../../config/locations";
+import { E_DB_WRITE } from "../../../config/constants";
 
 export default async function populateMenus() {
     const date = new Date();
